Refactor scanner logic into dedicated module
